rename package to secret
