+++ service/controller/resource/secret/create.go
@@ -1,4 +1,4 @@
-package namespace
+package secret
 
 import (
 	"context"
