+++ service/controller/resource/secret/delete.go
@@ -1,4 +1,4 @@
-package namespace
+package secret
 
 import (
 	"context"
